Move Action Copy to Isolated Module
